+++ src/main/java/net/pincette/json/streams/KafkaProvider.java
@@ -28,7 +28,6 @@ import static org.apache.kafka.clients.consumer.ConsumerConfig.VALUE_DESERIALIZE
 
 import com.typesafe.config.Config;
 import java.time.Duration;
-import java.util.Collections;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;