+++ src/main/java/net/pincette/json/streams/S3AttachmentsStage.java
@@ -10,20 +10,18 @@ import static java.util.UUID.randomUUID;
 import static java.util.concurrent.CompletableFuture.completedFuture;
 import static java.util.stream.Collectors.joining;
 import static java.util.stream.Stream.concat;
-import static javax.net.ssl.KeyManagerFactory.getDefaultAlgorithm;
 import static net.pincette.json.JsonUtil.arrayValue;
-import static net.pincette.json.JsonUtil.asArray;
 import static net.pincette.json.JsonUtil.asString;
 import static net.pincette.json.JsonUtil.createObjectBuilder;
 import static net.pincette.json.JsonUtil.emptyArray;
 import static net.pincette.json.JsonUtil.emptyObject;
 import static net.pincette.json.JsonUtil.getValue;
-import static net.pincette.json.JsonUtil.isArray;
 import static net.pincette.json.JsonUtil.isObject;
 import static net.pincette.json.JsonUtil.isString;
 import static net.pincette.json.JsonUtil.objectValue;
 import static net.pincette.json.JsonUtil.stringValue;
 import static net.pincette.json.streams.Common.S3ATTACHMENTS;
+import static net.pincette.json.streams.Common.getRequestBuilder;
 import static net.pincette.json.streams.Common.tryToGetForever;
 import static net.pincette.json.streams.Logging.logStageObject;
 import static net.pincette.json.streams.S3Util.getObject;
@@ -33,15 +31,11 @@ import static net.pincette.rs.Base64.base64Encoder;
 import static net.pincette.rs.Chain.with;
 import static net.pincette.rs.PassThrough.passThrough;
 import static net.pincette.util.Builder.create;
-import static net.pincette.util.Pair.pair;
 import static net.pincette.util.StreamUtil.composeAsyncStream;
 import static net.pincette.util.Util.getLastSegment;
 import static net.pincette.util.Util.must;
 import static net.pincette.util.Util.segments;
-import static net.pincette.util.Util.tryToDoRethrow;
-import static net.pincette.util.Util.tryToGetRethrow;
 
-import java.io.FileInputStream;
 import java.net.URI;
 import java.net.http.HttpClient;
 import java.net.http.HttpClient.Redirect;
@@ -49,7 +43,6 @@ import java.net.http.HttpClient.Version;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
 import java.nio.ByteBuffer;
-import java.security.KeyStore;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -63,10 +56,6 @@ import java.util.stream.Stream;
 import javax.json.JsonArray;
 import javax.json.JsonObject;
 import javax.json.JsonValue;
-import javax.net.ssl.KeyManager;
-import javax.net.ssl.KeyManagerFactory;
-import javax.net.ssl.SSLContext;
-import net.pincette.function.SideEffect;
 import net.pincette.json.JsonUtil;
 import net.pincette.mongo.streams.Stage;
 import net.pincette.rs.Concat;
@@ -84,9 +73,7 @@ class S3AttachmentsStage {
   private static final String HEADERS = "headers";
   private static final String HTTP_ERROR = "httpError";
   private static final String KEY = "key";
-  private static final String KEY_STORE = "keyStore";
   private static final String MIME_VERSION = "MIME-Version";
-  private static final String PASSWORD = "password";
   private static final String SSL_CONTEXT = "sslContext";
   private static final String STATUS_CODE = "statusCode";
   private static final String URL = "url";
@@ -182,23 +169,6 @@ class S3AttachmentsStage {
                     .thenApply(publisher -> request.POST(fromPublisher(publisher)).build()));
   }
 
-  private static Optional<SSLContext> createSslContext(final JsonObject sslContext) {
-    final String password = sslContext.getString(PASSWORD);
-
-    return tryToGetRethrow(() -> SSLContext.getInstance("TLSv1.3"))
-        .flatMap(
-            context ->
-                getKeyStore(sslContext.getString(KEY_STORE), password)
-                    .flatMap(store -> getKeyManagerFactory(store, password))
-                    .flatMap(S3AttachmentsStage::getKeyManagers)
-                    .map(managers -> pair(context, managers)))
-        .map(
-            pair ->
-                SideEffect.<SSLContext>run(
-                        () -> tryToDoRethrow(() -> pair.first.init(pair.second, null, null)))
-                    .andThenGet(() -> pair.first));
-  }
-
   private static Function<JsonObject, CompletionStage<JsonObject>> execute(
       final HttpClient client,
       final Function<JsonObject, JsonValue> url,
@@ -237,37 +207,10 @@ class S3AttachmentsStage {
         newBuilder().version(Version.HTTP_1_1).followRedirects(Redirect.NORMAL);
 
     return ofNullable(expression.getJsonObject(SSL_CONTEXT))
-        .flatMap(S3AttachmentsStage::createSslContext)
+        .flatMap(Common::createSslContext)
         .map(builder::sslContext)
         .orElse(builder)
         .build();
-  }
-
-  private static Optional<KeyManager[]> getKeyManagers(final KeyManagerFactory factory) {
-    return Optional.of(factory.getKeyManagers()).filter(managers -> managers.length > 0);
-  }
-
-  private static Optional<KeyManagerFactory> getKeyManagerFactory(
-      final KeyStore keyStore, final String password) {
-    return tryToGetRethrow(() -> KeyManagerFactory.getInstance(getDefaultAlgorithm()))
-        .map(
-            factory ->
-                SideEffect.<KeyManagerFactory>run(
-                        () -> tryToDoRethrow(() -> factory.init(keyStore, password.toCharArray())))
-                    .andThenGet(() -> factory));
-  }
-
-  private static Optional<KeyStore> getKeyStore(final String keyStore, final String password) {
-    return tryToGetRethrow(() -> KeyStore.getInstance("pkcs12"))
-        .map(
-            store ->
-                SideEffect.<KeyStore>run(
-                        () ->
-                            tryToDoRethrow(
-                                () ->
-                                    store.load(
-                                        new FileInputStream(keyStore), password.toCharArray())))
-                    .andThenGet(() -> store));
   }
 
   private static Publisher<ByteBuffer> headerPublisher(
@@ -308,18 +251,7 @@ class S3AttachmentsStage {
 
   private static HttpRequest.Builder setHeaders(
       final HttpRequest.Builder builder, final JsonObject headers) {
-    return concat(
-            headers.entrySet().stream()
-                .filter(e -> isString(e.getValue()))
-                .map(e -> pair(e.getKey(), asString(e.getValue()).getString())),
-            headers.entrySet().stream()
-                .filter(e -> isArray(e.getValue()))
-                .flatMap(
-                    e ->
-                        asArray(e.getValue()).stream()
-                            .flatMap(v -> stringValue(v).stream())
-                            .map(s -> pair(e.getKey(), s))))
-        .reduce(builder, (b, p) -> b.header(p.first, p.second), (b1, b2) -> b1);
+    return getRequestBuilder(builder, headers);
   }
 
   static Stage s3AttachmentsStage(final Context context) {